funciones en login sacadas al gestorDeUsuarios para respetar el MVC
